More appropriate stack digestion

+++ code/modules/vore/eating/digest_act_vr.dm
@@ -49,7 +49,15 @@
 					qdel(O)
 			else if(item_storage)
 				O.forceMove(item_storage)
-		qdel(src)
+		if(istype(src,/obj/item/stack))
+			var/obj/item/stack/S = src
+			if(S.get_amount() <= 1)
+				qdel(src)
+			else
+				S.use(1)
+				digest_stage = w_class
+		else
+			qdel(src)
 	if(g_damage > w_class)
 		return w_class
 	return g_damage
